backend/requirements.txt
fastapi[standard]>=0.104.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain>=0.1.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
ruff>=0.1.0
mypy>=1.7.0
black>=23.0.0